+++ include/tools/types/sfinae/available-2013.hpp
@@ -16,14 +16,6 @@
 
 namespace tools     {
 namespace sfinae    {
-
-    template<class a, class b>
-    struct decltype_ 
-    { 
-        using type = ::std::true_type; 
-        using second = b;
-    };
-
 namespace available {
 
     namespace detail
@@ -134,34 +126,6 @@ namespace available {
             begin_() = delete;
             enum { value = result_t::value };
         };
-
-#if 0
-        // diagnostic version
-        template<class t> class begin_
-        {
-        public:
-            using x = ::std::remove_reference_t<t>;
-
-            template <class u> static 
-                decltype_<u, decltype(::std::declval<u>().begin())>
-				check(u*);
-
-            template <class> static
-                ::std::false_type check(...);
-
-            using check_t = decltype(check<x>(nullptr));
-
-            enum { v = ! ::std::is_same<check_t, ::std::false_type>::value };
-
-            using cond_t
-                = ::std::conditional<v, ::std::true_type, ::std::false_type>;
-
-            using result_t = typename cond_t::type;
-        public:
-            begin_() = delete;
-            enum { value = result_t::value };
-        };
-#endif
 
     } // namespace detail
